Add option to customize spacing applied to SwiftUI-based paywall footer
This patch makes it possible for API users to customize the amount
of spacing that's applied to a `PaywallView` when presented as a
footer using the `paywallFooter` SwiftUI modifier. That API uses
SwiftUI's `safeAreaInset` modifier internally, which by default
adds a system-defined amount of spacing between the view placed
within the safe area inset and the host view that the modifier is
applied to. With this change, API users are now free to remove or
modify that spacing, which is required for certain app designs.

This is a backward-compatible change, as the added `spacing`
parameter is optional with a default value of `nil`, which retains
the existing behavior when the new parameter isn't passed by the
API user.