When taking screenshots for validation, ignore safe area. (#5376)
* When taking screenshots for validation, ignore safe area.

This fixes an issue where screenshots were being cutoff on the bottom on Mac Catalyst, I believe because the view was being pushed down to insert a window titlebar.

We should be careful to separately verify that the presentation of paywalls is sized correctly, but I think we can't verify that through these screenshots, we need to use the actual mechanisms for presenting paywalls in sheets, etc.

* Revert the change in iOS deployment target, and instead just mark the extension as only avaliable on the same OS versions as TakeScreenshotTest.

---------

Co-authored-by: Josh Holtz <me@joshholtz.com>