Video Component Views (#5527)
This adds a video component for apple platforms: iOS, macOS, and watchOS. There is are differences in them. But the primary difference that is noticeable in the UX is that iOS is the only one that does not have a black background if the video aspect ratio is not the same as the view's and does not fill its size.

> [!NOTE]
> macOS and watchOS will ship with the standard black background on the video component. I don't have a good solution for handling that right now.